feat: refactor imports and create query client utility for better organization
