+++ frontend/src/main.tsx
@@ -1,12 +1,9 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import App from './App.jsx'
-import { BrowserRouter } from 'react-router'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-
-
-export const queryClient = new QueryClient();
+import App from '@/App'
+import { QueryClientProvider } from '@tanstack/react-query'
+import { queryClient } from './utils/query-client'
 
 createRoot(document.getElementById('root') as HTMLElement).render(
   <StrictMode>
