+++ frontend/src/utils/query-client.ts
@@ -1,0 +1,2 @@
+import { QueryClient } from '@tanstack/react-query'
+export const queryClient = new QueryClient();
